move identifier for the same module as evidence metadata
