+++ test/schema/test_value_schema.py
@@ -8,8 +8,7 @@ import pytest
 
 from mlte._private.schema import validate_value_schema
 from mlte.api import read_value
-from mlte.evidence.identifier import Identifier
-from mlte.evidence.metadata import EvidenceMetadata
+from mlte.evidence.metadata import EvidenceMetadata, Identifier
 from mlte.value.types.integer import Integer
 from mlte.value.types.opaque import Opaque
 from mlte.value.types.real import Real
